Rename ThemeSwitch component
